added wind_model to config to distinguish between the models, gridded data was being tested twice and implemented wind ninja
