+++ smrf/framework/CoreConfig.ini
@@ -425,6 +425,11 @@ description = Determines if the x and y coordinates are interpreted as on a
 
 [wind]
 
+wind_model:
+default = winstral,
+options = [winstral wind_ninja interp],
+description = Wind model to interpolate wind measurements to the model domain
+
 distribution:
 default = idw,
 options = [dk idw grid kriging],
