+++ tests/Lakes/config.ini
@@ -59,6 +59,7 @@ grid_local:                    True
 ################################################################################
 
 [wind]
+wind_model:                    wind_ninja
 distribution:                  grid
 wind_ninja_dir:                ./input
 wind_ninja_dxdy:               50
