+++ tests/test_load_data.py
@@ -16,9 +16,6 @@ from inicheck.tools import cast_all_variables
 from smrf.framework.model_framework import can_i_run_smrf, run_smrf
 from tests.nwrc_check import NWRCCheck
 from tests.smrf_test_case import SMRFTestCase
-import smrf
-from os.path import join, dirname, abspath
-import numpy as np
 
 
 @unittest.skipUnless(
@@ -135,174 +132,6 @@ class TestLoadCSVData(SMRFTestCase):
         # apply the new recipies
         config.apply_recipes()
         config = cast_all_variables(config, config.mcfg)
-
-        self.assertIsNone(run_smrf(config))
-
-
-class TestLoadGrid(SMRFTestCase):
-
-    def test_grid_wrf(self):
-        """ WRF NetCDF loading """
-
-        config = deepcopy(self.base_config)
-        del config.raw_cfg['csv']
-
-        adj_config = {'gridded': {'data_type': 'wrf',
-                                  'wrf_file': './RME/gridded/WRF_test.nc',
-                                  'northern_hemisphere':'True'},
-                      'system': {'threading': 'False',
-                                 'log_file': './output/log.txt'},
-                      'precip': {'station_adjust_for_undercatch': 'False'},
-                      'thermal': {'correct_cloud': 'False',
-                                  'correct_veg': 'True'},
-                      'time': {'start_date': '2015-03-03 00:00',
-                               'end_date': '2015-03-03 04:00'}}
-
-        config.raw_cfg.update(adj_config)
-
-        # set the distribution to grid, thermal defaults will be fine
-        for v in SMRFTestCase.dist_variables:
-            config.raw_cfg[v]['grid_mask'] = 'False'
-
-        # fix the time to that of the WRF_test.nc
-        config.apply_recipes()
-        config = cast_all_variables(config, config.mcfg)
-
-        # ensure that the recipes are used
-        assert 'station_adjust_for_undercatch' not in config.cfg['precip'].keys()
-        self.assertFalse(config.cfg['thermal']['correct_cloud'])
-        self.assertTrue(config.cfg['thermal']['correct_veg'])
-
-        self.assertIsNone(run_smrf(config))
-
-    def test_grid_hrrr(self):
-        """ HRRR grib2 loading """
-
-        config = deepcopy(self.base_config)
-        del config.raw_cfg['csv']
-
-        hrrr_grid = {'data_type': 'hrrr_grib',
-                     'hrrr_directory': './RME/gridded/hrrr_test/',
-                     'northern_hemisphere':'True'}
-        config.raw_cfg['gridded'] = hrrr_grid
-        config.raw_cfg['system']['threading'] = 'False'
-
-        # set the distribution to grid, thermal defaults will be fine
-        for v in SMRFTestCase.dist_variables:
-            config.raw_cfg[v]['distribution'] = 'grid'
-            config.raw_cfg[v]['grid_mask'] = 'False'
-
-        config.raw_cfg['precip']['station_adjust_for_undercatch'] = 'False'
-        config.raw_cfg['thermal']['correct_cloud'] = 'True'
-        config.raw_cfg['thermal']['correct_veg'] = 'True'
-
-        # fix the time to that of the WRF_test.nc
-        config.raw_cfg['time']['start_date'] = '2018-07-22 16:00'
-        config.raw_cfg['time']['end_date'] = '2018-07-22 20:00'
-
-        #config.raw_cfg['system']['log_file'] = 'none'
-
-        config.apply_recipes()
-        config = cast_all_variables(config, config.mcfg)
-
-        # ensure that the recipes are used
-        result = (
-            'station_adjust_for_undercatch' not in config.cfg['precip'].keys()
-        )
-        self.assertTrue(result)
-        self.assertTrue(config.cfg['thermal']['correct_cloud'])
-        self.assertTrue(config.cfg['thermal']['correct_veg'])
-
-        self.assertIsNone(run_smrf(config))
-
-    def test_grid_hrrr_local(self):
-        """ HRRR grib2 loading with local elevation gradient """
-
-        config = deepcopy(self.base_config)
-        del config.raw_cfg['csv']
-
-        hrrr_grid = {'data_type': 'hrrr_grib',
-                     'hrrr_directory': './RME/gridded/hrrr_test/',
-                     'northern_hemisphere':'True'}
-        config.raw_cfg['gridded'] = hrrr_grid
-        config.raw_cfg['system']['threading'] = 'False'
-        config.raw_cfg['system']['log_file'] = './output/log.txt'
-
-        # set the distribution to grid, thermal defaults will be fine
-        for v in SMRFTestCase.dist_variables:
-            config.raw_cfg[v]['distribution'] = 'grid'
-            config.raw_cfg[v]['grid_mask'] = 'False'
-
-        # local gradient
-        config.raw_cfg['air_temp']['grid_local'] = 'True'
-        config.raw_cfg['air_temp']['grid_local_n'] = '25' # only 47 grid cells
-
-        config.raw_cfg['vapor_pressure']['grid_local'] = 'True'
-
-        # Only 47 grid cells in domain
-        config.raw_cfg['vapor_pressure']['grid_local_n'] = '25'
-        config.raw_cfg['precip']['grid_local'] = 'True'
-        config.raw_cfg['precip']['grid_local_n'] = '25'
-        config.raw_cfg['thermal']['correct_cloud'] = 'True'
-        config.raw_cfg['thermal']['correct_veg'] = 'True'
-
-        # fix the time to that of the WRF_test.nc
-        config.raw_cfg['time']['start_date'] = '2018-07-22 16:00'
-        config.raw_cfg['time']['end_date'] = '2018-07-22 20:00'
-
-        config.apply_recipes()
-        config = cast_all_variables(config, config.mcfg)
-
-        # ensure that the recipes are used
-        self.assertTrue(
-            'station_adjust_for_undercatch' not in config.cfg['precip'].keys()
-        )
-        self.assertTrue(config.cfg['thermal']['correct_cloud'])
-        self.assertTrue(config.cfg['thermal']['correct_veg'])
-
-        self.assertIsNone(run_smrf(config))
-
-    def test_grid_netcdf(self):
-        """ Generic NetCDF loading """
-        config = deepcopy(self.base_config)
-        del config.raw_cfg['csv']
-
-        generic_grid = {'data_type': 'netcdf',
-                        'netcdf_file': './RME/gridded/netcdf_test.nc',
-                        'northern_hemisphere':'True',
-                        'air_temp': 'air_temp',
-                        'vapor_pressure': 'vapor_pressure',
-                        'precip': 'precip',
-                        'wind_speed': 'wind_speed',
-                        'wind_direction': 'wind_direction',
-                        'thermal': 'thermal',
-                        'cloud_factor': 'cloud_factor'}
-        config.raw_cfg['gridded'] = generic_grid
-        config.raw_cfg['system']['time_out'] = '25'
-        config.raw_cfg['system']['queue_max_values'] = '2'
-        config.raw_cfg['system']['threading'] = 'False' # Doesn't work with true
-
-        # set the distribution to grid, thermal defaults will be fine
-        for v in SMRFTestCase.dist_variables:
-            config.raw_cfg[v]['distribution'] = 'grid'
-            config.raw_cfg[v]['grid_mask'] = 'False'
-
-        config.raw_cfg['thermal']['correct_cloud'] = 'False'
-        config.raw_cfg['thermal']['correct_veg'] = 'True'
-
-        # fix the time to that of the WRF_test.nc
-        config.raw_cfg['time']['start_date'] = '2015-03-03 00:00'
-        config.raw_cfg['time']['end_date'] = '2015-03-03 04:00'
-
-        config.apply_recipes()
-        config = cast_all_variables(config, config.mcfg)
-
-        # ensure that the recipes are used
-        self.assertTrue(
-            'station_adjust_for_undercatch' not in config.cfg['precip'].keys()
-        )
-        self.assertFalse(config.cfg['thermal']['correct_cloud'])
-        self.assertTrue(config.cfg['thermal']['correct_veg'])
 
         self.assertIsNone(run_smrf(config))
